storage/fs: move {Create,Record,Cleanup}TempDir from storage
Move `CreateTempDir`, `RecordTempDir` and `CleanupTempDirs` functions
from the storage package into fs. They specifically relate to the
filesystem and can be moved to help reduce the storage package's surface
area.

Release note: None
