+++ pkg/storage/BUILD.bazel
@@ -34,7 +34,6 @@ go_library(
         "slice_go1.9.go",
         "sst_iterator.go",
         "sst_writer.go",
-        "temp_dir.go",
         "temp_engine.go",
         "testing_knobs.go",
         ":gen-resourcelimitreached-stringer",  # keep
@@ -64,7 +63,6 @@ go_library(
         "//pkg/util/log",
         "//pkg/util/mon",
         "//pkg/util/protoutil",
-        "//pkg/util/stop",
         "//pkg/util/syncutil",
         "//pkg/util/sysutil",
         "//pkg/util/timeutil",
@@ -112,7 +110,6 @@ go_test(
         "resource_limiter_test.go",
         "sst_iterator_test.go",
         "sst_writer_test.go",
-        "temp_dir_test.go",
         "temp_engine_test.go",
     ],
     data = glob(["testdata/**"]),
@@ -142,7 +139,6 @@ go_test(
         "//pkg/util/protoutil",
         "//pkg/util/randutil",
         "//pkg/util/shuffle",
-        "//pkg/util/stop",
         "//pkg/util/sysutil",
         "//pkg/util/timeutil",
         "//pkg/util/uint128",
