+++ pkg/storage/fs/temp_dir_test.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package storage
+package fs
 
 import (
 	"bytes"
